+++ guidata/dataset/__init__.py
@@ -1,26 +1,24 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2009-2010 CEA
-# Pierre Raybaut
-# Licensed under the terms of the CECILL License
-# (see guidata/__init__.py for details)
+# Licensed under the terms of the BSD 3-Clause
+# (see guidata/LICENSE for details)
 
 """
 dataset
 =======
 
-The ``guidata.dataset`` package provides the core features for data set 
+The ``guidata.dataset`` package provides the core features for data set
 display and editing with automatically generated graphical user interfaces.
 
 .. automodule:: guidata.dataset.dataitems
    :members:
-       
+
 .. automodule:: guidata.dataset.datatypes
    :members:
-       
+
 .. automodule:: guidata.dataset.qtitemwidgets
    :members:
-       
+
 .. automodule:: guidata.dataset.qtwidgets
    :members:
 """