+++ guidata/env.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
-# Licensed under the terms of the BSD 3-Clause or the CeCILL-B License
-# (see guidata/__init__.py for details)
+# Licensed under the terms of the BSD 3-Clause
+# (see guidata/LICENSE for details)
 
 """
 Execution environmnent utilities