+++ guidata/tests/test_bool_selector.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2009-2010 CEA
-# Pierre Raybaut
-# Licensed under the terms of the CECILL License
-# (see guidata/__init__.py for details)
+# Licensed under the terms of the BSD 3-Clause
+# (see guidata/LICENSE for details)
 
 """
 DataItem groups and group selection