+++ guidata/utils/gettext_helpers.py
@@ -1,14 +1,12 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2009-2010 CEA
-# Pierre Raybaut
-# Licensed under the terms of the CECILL License
-# (see guidata/__init__.py for details)
+# Licensed under the terms of the BSD 3-Clause
+# (see guidata/LICENSE for details)
 
-import sys
 import os
 import os.path as osp
 import subprocess
+import sys
 
 if os.name == "nt":
     # Find pygettext.py source on a windows install